feat(PMInstaller): don't relaunch install prompts until result received
This was supposed to fix one type of ROM bug where clicking outside of the install prompt would never return a result to the application. However, the workaround was to keep relaunching the install prompt until the user either actually clicked install or cancel. This apparently is also incorrectly handled by a lot of ROMs and causes the install prompt to keep relaunching, even after pressing the install button.

I don't think there's a good solution, so I'll just remove the relaunching behavior since it's more annoying to the user.